pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "dranspose"
authors = [
  { name="Felix Engelmann", email="felix-github@nlogn.org" },
]
description = "A stream matrix transposition framework"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
license = "MIT"
dependencies = [
    "pyzmq",
    "redis[hiredis]",
    "fastapi",
    "uvicorn",
    "numpy",
    # "requests",
    "pydantic",
    "pydantic-settings",
    # limit cbor until https://github.com/agronholm/cbor2/issues/272 is resolved.
    "cbor2 < 5.8",
    "redis-logs",
    "h5py",
]
dynamic = ["version"]

[tool.setuptools.packages.find]
include = ["dranspose*"]  # package names should match these glob patterns (["*"] by default)
exclude = ["tests*"]  # exclude packages matching these glob patterns (empty by default)
#namespaces = false

[tool.setuptools.package-data]
dranspose = ["frontend/*.html"]

[project.optional-dependencies]
tests = [
    "pytest",
    "pytest-asyncio",
    "aiohttp",
    "psutil >= 7.2.1",
    "types-psutil",
    "pytest-plt",
    "h5pyd",
    "readerwriterlock",
]
docs = [
    "mkdocs",
    "mkdocstrings[python]",
    "mkdocs-material",
    "mkdocs-swagger-ui-tag"
]
bench = [
    "psutil >= 7.2.1",
    "types-psutil",
]

[project.scripts]
dranspose = "dranspose.cli:run"

[tool.setuptools_scm]

[project.urls]
"Homepage" = "https://github.com/felix-engelmann/dranspose"
"Bug Tracker" = "https://github.com/felix-engelmann/dranspose/issues"


[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
addopts = "--strict-markers"
markers = [
    "allow_errors_in_log: do not fail test if there are ERROR logging messages",
]

[tool.mypy]
plugins = [
  "pydantic.mypy"
]

follow_imports = "silent"
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true

# for strict mypy: (this is the tricky one :-))
disallow_untyped_defs = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[[tool.mypy.overrides]]
module = ["h5pyd.*", "rlh.*"]
ignore_missing_imports = true

[tool.black]
extend-exclude = '''
# A regex preceded with ^/ will apply only to files and directories
# in the root of the project.
(
  docs/*
)
'''
